+++ app/block/publish_event.go
@@ -49,7 +49,7 @@ func PublishEvent(blockNumber uint64, event *db.Events, redis *d.RedisInfo) bool
 		BlockHash:       event.BlockHash,
 	}
 
-	if err := redis.Client.Publish(context.Background(), "event", data).Err(); err != nil {
+	if err := redis.Client.Publish(context.Background(), redis.EventPublishTopic, data).Err(); err != nil {
 
 		log.Printf("❗️ Failed to publish event from block %d : %s\n", blockNumber, err.Error())
 		return false
